app status modal abstraction
