--- src/components/app/details/appDetails/deploymentStatusModal.css
@@ -1,8 +1,0 @@
-
-.deployment-status-modal {
-    position: fixed;
-    width: 800px;
-    right: 0;
-    top: 0;
-    height: 100%;
-}
